--- src/routes/Game/styles.module.css
@@ -1,40 +1,0 @@
-.root {
-    margin: 0 auto;
-}
-.container {
-    width: fit-content;
-    position: absolute;
-    right: 0;
-    top: 50%;
-    left: 50%;
-    opacity: 0;
-    transform: translate(-50%, -50%);
-    animation: text-fade-in 800ms 600ms forwards;
-}
-
-@keyframes text-fade-in {
-    from {
-        right: 0;
-    }
-    to {
-        opacity: 1;
-        right: 25%;
-    }
-}
-
-.buttons {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: center;
-}
-.button {
-    margin: 0 10px;
-}
-
-.flex {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-wrap: wrap;
-}
